requirements.txt
annotated-types==0.5.0
anyio==3.7.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.103.2
fastapi-cli==0.0.1
greenlet==3.1.1
idna==3.10
importlib-metadata==1.7.0
passlib==1.7.4
pyasn1==0.4.8
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==0.21.1
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.27.0
typing-extensions==4.7.1
zipp==3.15.0
